# sez/src/main/resources/api/swagger_full.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "API for managing the tutoring system (CRM)",
    "title" : "SEZ Tutoring Platform API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local Server",
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
...
    "/api/v1/tutors/{id}" : {
      "get" : {
        "operationId" : "getTutorById",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TutorResponse"
                }
              }
            },
            "description" : "Tutor profile retrieved successfully"
          },
          "404" : {
            "description" : "Tutor not found"
          }
        },
        "summary" : "Get tutor profile by ID",
        "tags" : [ "Tutors" ]
      },
      "patch" : {
        "operationId" : "updateTutor",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTutorDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Tutor profile updated successfully"
          },
          "404" : {
            "description" : "Tutor not found"
          }
        },
        "summary" : "Update tutor profile data",
        "tags" : [ "Tutors" ]
      }
    },
    "/api/v1/students" : {
      "get" : {
        "operationId" : "getStudents",
        "parameters" : [ {
          "description" : "Filter by student name or surname",
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Filter by course (subject) name",
          "in" : "query",
          "name" : "courseName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/StudentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of students retrieved successfully"
          }
        },
        "summary" : "Get list of students for current tutor",
        "tags" : [ "Students" ]
      }
    },
    "/api/v1/students/{id}" : {
      "get" : {
        "operationId" : "getStudentById",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StudentResponse"
                }
              }
            },
            "description" : "Student data retrieved successfully"
          },
          "404" : {
            "description" : "Student not found"
          }
        },
        "summary" : "Get student by ID",
        "tags" : [ "Students" ]
      },
      "patch" : {
        "operationId" : "updateStudent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateStudentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Student profile updated successfully"
          },
          "404" : {
            "description" : "Student not found"
          }
        },
        "summary" : "Update student profile data",
        "tags" : [ "Students" ]
      }
    },
    "/api/v1/courses/{courseId}/students" : {
      "get" : {
        "operationId" : "getStudentsByCourseId",
        "parameters" : [ {
          "in" : "path",
          "name" : "courseId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/StudentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of students by course retrieved successfully"
          },
          "404" : {
            "description" : "Course not found"
          }
        },
        "summary" : "Get list of students enrolled on the course",
        "tags" : [ "Students" ]
      }
    },
    "/api/v1/subjects" : {
      "get" : {
        "operationId" : "getSubjects",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SubjectResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of subjects retrieved successfully"
          }
        },
        "summary" : "Get list of all available subjects",
        "tags" : [ "Subjects" ]
      },
      "post" : {
        "operationId" : "createSubject",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSubjectDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Subject created successfully"
          }
        },
        "summary" : "Create a new subject",
        "tags" : [ "Subjects" ]
      }
    },
    "/api/v1/subjects/{id}" : {
      "delete" : {
        "description" : "Deletion will be blocked if any courses are linked to this subject",
        "operationId" : "deleteSubject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Subject deleted successfully"
          },
          "400" : {
            "description" : "Cannot delete subject linked to existing courses"
          },
          "404" : {
            "description" : "Subject not found"
          }
        },
        "summary" : "Delete subject by ID",
        "tags" : [ "Subjects" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TutorResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "patronymic" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "hourlyRate" : {
            "format" : "double",
            "type" : "number"
          },
          "rating" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "required" : [ "bio", "email", "hourlyRate", "id", "name", "rating", "role", "surname" ],
        "type" : "object"
      },
      "UpdateTutorDto" : {
        "description" : "Fields that can be updated in tutor profile",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "patronymic" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "hourlyRate" : {
            "type" : "number"
          }
        },
        "required" : [ "bio", "hourlyRate", "name", "surname" ],
        "type" : "object"
      },
      "StudentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "role" : {
            "enum" : [ "STUDENT" ],
            "type" : "string"
          }
        },
        "required" : [ "email", "id", "name", "role", "surname" ],
        "type" : "object"
      },
      "UpdateStudentDto" : {
        "properties" : {
          "name" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "surname" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          },
          "patronymic" : {
            "maxLength" : 50,
            "minLength" : 2,
            "type" : "string"
          }
        },
        "required" : [ "firstName", "lastName" ],
        "type" : "object"
      },
      "SubjectResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreateSubjectDto" : {
        "properties" : {
          "name" : {
            "description" : "Name of the subject (e.g., Mathematics, English)",
            "maxLength" : 100,
            "minLength" : 2,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
    }
  }
}